backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM and Agent Framework
langchain==0.1.5
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph==0.0.20
openai==1.10.0
anthropic==0.17.0

# Vision Processing
pillow==10.2.0
opencv-python==4.9.0.80

# Speech Processing
openai-whisper==20231117
deepgram-sdk==3.2.0
soundfile==0.12.1
librosa==0.10.1

# Vector Database
chromadb==0.4.22
faiss-cpu==1.7.4
sentence-transformers==2.3.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Data Processing
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# Caching and Performance
diskcache==5.6.3
aiohttp==3.9.3
httpx==0.26.0

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.1
tenacity==8.2.3
pyyaml==6.0.1

# Type checking
mypy==1.8.0
types-pyyaml==6.0.12.12
